\name{allmypackages-package}
\alias{allmypackages-package}
\alias{allmypackages}
\docType{package}
\title{
\packageTitle{allmypackages}
}
\description{
\packageDescription{allmypackages}
}
\details{
Does nothing.
}
\author{
\packageAuthor{allmypackages}

Maintainer: \packageMaintainer{allmypackages}
}

\keyword{package}

